Split render engines into separate files and add unit tests
