+++ raysect/core/workflow/tests/meson.build
@@ -1,12 +1,12 @@
 # WARNING: This file is automatically generated by dev/generate_meson_files.py.
 # The template file used to generate this file is dev/subdir-meson.build.
 
-target_path = 'raysect/core'
+target_path = 'raysect/core/workflow/tests'
 
 # source files
-py_files = ['__init__.py', 'constants.py']
-pyx_files = ['boundingbox.pyx', 'boundingsphere.pyx', 'containers.pyx', 'intersection.pyx', 'material.pyx', 'ray.pyx']
-pxd_files = ['__init__.pxd', 'boundingbox.pxd', 'boundingsphere.pxd', 'containers.pxd', 'intersection.pxd', 'material.pxd', 'ray.pxd']
+py_files = ['__init__.py', 'test_workflow.py']
+pyx_files = []
+pxd_files = []
 data_files = []
 
 # compile cython
@@ -27,8 +27,3 @@ py.install_sources(
     subdir: target_path
 )
 
-subdir('acceleration')
-subdir('math')
-subdir('scenegraph')
-subdir('tests')
-subdir('workflow')
